backend/requirements.txt
alembic==1.18.1
APScheduler==3.11.2
bcrypt==5.0.0
blinker==1.9.0
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
Flask==3.1.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.0
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
mysqlclient==2.2.8
packaging==26.0
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.11.0
pymongo==4.16.0
pytest==9.0.2
python-dotenv==1.2.1
SQLAlchemy==2.0.46
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
Werkzeug==3.1.5
pytest-mock==3.14.0
mongomock==4.1.2